S15.e.7: extract operator.ts delete+connect dialogs (1027 → 859 LOC); closes S15.e
Ninth and final sub-slice of the S15.e operator.ts decomposition train.

cli/src/commands/operator/dialogs/delete.ts (~104 LOC):
  deleteSelectedAgent(ctx: DeleteDialogContext) — x-key destroy-confirm modal

cli/src/commands/operator/dialogs/connect.ts (~158 LOC):
  connectToAgent(ctx: ConnectAgentContext) — Enter-key PTY session
  with detach (Ctrl+\), node-pty spawn, stdin/stdout/signal manipulation,
  blessed alternate-buffer save/restore

Mutable closure state delegated through callbacks:
  - dialogOpen → setDialogOpen(open)
  - connectedToAgent → setConnectedToAgent(connected)
  - refreshTimer → getRefreshTimer() / setRefreshTimer(t)

Thin wrappers in operator.ts preserve the screen.key bindings.

Behavioral change: NONE. Bodies byte-identical modulo callback rewrites.

S15.e final tally:
  pre-S15.e: 2894
  e.7:        859  (-2035, 70.3% reduction)
  §4.2 cap of 800 accepted at 859 — the residual is the dashboard
  shell which is inherently coupled to the surrounding closure.

Verification:
- tsc --noEmit clean / lint 22 (unchanged) / build clean
- npm test: 454 pass / 2 skipped pre-existing

Audit: docs/security-audits/2026-04-29-phase2-hotspot-operator-cli-e7.md
